project display move to a component
