+++ app/portafolio/portafolio.module.css
@@ -1,4 +1,5 @@
 .main {
+  padding-top: 120px;
   min-height: 100svh;
 }
 
@@ -17,27 +18,6 @@
   width: 100vw;
   height: 100%;
   z-index: -1;
-}
-
-.appContainer {
-  display: flex;
-}
-
-.textSection {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  padding: 40px 0;
-  padding-right: 100px;
-  /* padding-left: 120px; */
-}
-
-.image {
-  object-fit: cover;
-  width: 100%;
-  height: auto;
 }
 
 /* .section {
