+++ components/projectDisplay/projectDisplay.module.css
@@ -1,0 +1,20 @@
+.appContainer {
+  display: flex;
+}
+
+.textSection {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  padding: 40px 0;
+  padding-right: 100px;
+  /* padding-left: 120px; */
+}
+
+.image {
+  object-fit: cover;
+  width: 100%;
+  height: auto;
+}
