+++ data/projects.json
@@ -14,10 +14,10 @@
     "id": 2,
     "name": "Ticket Administration System",
     "technologies": "React, Django Rest API, PostgreSQL",
-    "description": "Lorem ipsum lorem a multi-account system on AWS for a landing zone. Created and documented multiple Lambdas and alarms in CloudWatch to monitor the deployments of multiple projects Implemented automatiozations for a DevOps environment",
+    "description": "Lorem ipsum lorem a multi-account system on AWS for a landing zone. Created and documented multiple Lambdas and alarms in CloudWatch to monitor the deployments of multiple projects Implemented automatiozations for a DevOps environment Lorem ipsum lorem a multi-account system on AWS for a landing zone. Created and documented multiple Lambdas and alarms in CloudWatch to monitor the deployments of multiple projects Implemented automatiozations for a DevOps environment",
     "images": [
       {
-        "link": "https://drive.google.com/file/d/1XGBjzyKZsMRQTsK1F57qxr01zzXYPtx8/view?usp=sharing"
+        "link": "https://drive.google.com/file/d/1etrifNtiHVrSjH7SeVj3SHrraJktBsn9/view?usp=sharing"
       }
     ]
   },
@@ -28,7 +28,7 @@
     "description": "Lorem ipsum lorem a multi-account system on AWS for a landing zone. Created and documented multiple Lambdas and alarms in CloudWatch to monitor the deployments of multiple projects Implemented automatiozations for a DevOps environment",
     "images": [
       {
-        "link": "https://drive.google.com/file/d/1XGBjzyKZsMRQTsK1F57qxr01zzXYPtx8/view?usp=sharing"
+        "link": "https://drive.google.com/file/d/1ipG8p95CeW3o_VljhHZeJHbtCiMHlDYv/view?usp=sharing"
       }
     ]
   },
